requirements.txt
prefect==3.*
pandas==2.*
requests==2.*
pyyaml==6.*
SQLAlchemy==2.*
psycopg2-binary==2.*
python-dotenv==1.*
